Extension complete

+++ src/main/java/com/booleanuk/api/cinema/customers/CustomerController.java
@@ -1,14 +1,15 @@
 package com.booleanuk.api.cinema.customers;
 
-import org.apache.coyote.BadRequestException;
+import com.booleanuk.api.cinema.responses.CustomerListResponse;
+import com.booleanuk.api.cinema.responses.CustomerResponse;
+import com.booleanuk.api.cinema.responses.ErrorResponse;
+import com.booleanuk.api.cinema.responses.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @RestController
 @RequestMapping("customers")
@@ -16,54 +17,69 @@ public class CustomerController {
     @Autowired
     private CustomerRepository customerRepository;
 
+    private ErrorResponse errorResponse = new ErrorResponse();
+    private CustomerResponse customerResponse = new CustomerResponse();
+    private CustomerListResponse customerListResponse = new CustomerListResponse();
+
     @GetMapping
-    public ResponseEntity<List<Customer>> getAllCustomers() {
-        return ResponseEntity.ok(this.customerRepository.findAll());
+    public ResponseEntity<Response<?>> getAllCustomers() {
+        this.customerListResponse.set(this.customerRepository.findAll());
+        return ResponseEntity.ok(customerListResponse);
     }
 
     @PostMapping
-    public ResponseEntity<Customer> createCustomer(@RequestBody Customer customer) {
-        /*try {
-            customer.setCreatedAt(LocalDateTime.now());
-            customer.setUpdatedAt(LocalDateTime.now());
-            Customer customerToCreate = this.customerRepository.save(customer);
-            return new ResponseEntity<>(customerToCreate, HttpStatus.CREATED);
-        } catch (Exception e) {
-            throw new BadRequestException("Could not create a new customer, please check all fields are correct");
-            //return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }*/
+    public ResponseEntity<Response<?>> createCustomer(@RequestBody Customer customer) {
         customer.setCreatedAt(LocalDateTime.now());
         customer.setUpdatedAt(LocalDateTime.now());
-        return new ResponseEntity<Customer>(this.customerRepository.save(customer), HttpStatus.CREATED);
+        if (customer.getName() == null || customer.getEmail() == null || customer.getPhone() == null) {
+            this.errorResponse.set("Could not create a new customer, please check all fields are correct");
+            return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
+        }
+        this.customerResponse.set(this.customerRepository.save(customer));
+        return new ResponseEntity<>(customerResponse, HttpStatus.CREATED);
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<Customer> getCustomerById(@PathVariable int id) {
-        Customer customer = this.customerRepository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "No customer with that id found")
-        );
-        return ResponseEntity.ok(customer);
+    public ResponseEntity<Response<?>> getCustomerById(@PathVariable int id) {
+        Customer customer = this.customerRepository.findById(id).orElse(null);
+        if (customer == null) {
+            this.errorResponse.set("No customer with that id found");
+            return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+        }
+        this.customerResponse.set(customer);
+        return ResponseEntity.ok(customerResponse);
     }
 
     @PutMapping("{id}")
-    public ResponseEntity<Customer> updateCustomer(@PathVariable int id, @RequestBody Customer customer) {
-        Customer customerToUpdate = this.customerRepository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "No customer with that id was found")
-        );
+    public ResponseEntity<Response<?>> updateCustomer(@PathVariable int id, @RequestBody Customer customer) {
+        Customer customerToUpdate = this.customerRepository.findById(id).orElse(null);
+        if (customerToUpdate == null) {
+            this.errorResponse.set("No customer with that id found");
+            return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+        }
+
+        if (customer.getName() == null || customer.getEmail() == null || customer.getPhone() == null) {
+            this.errorResponse.set("Could not update the specified customer, please check all fields are correct");
+            return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
+        }
         customerToUpdate.setName(customer.getName());
         customerToUpdate.setEmail(customer.getEmail());
         customerToUpdate.setPhone(customer.getPhone());
         customerToUpdate.setUpdatedAt(LocalDateTime.now());
-        return new ResponseEntity<Customer>(this.customerRepository.save(customerToUpdate), HttpStatus.CREATED);
+        this.customerResponse.set(this.customerRepository.save(customerToUpdate));
+        return new ResponseEntity<>(customerResponse, HttpStatus.CREATED);
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<Customer> deleteCustomerById(@PathVariable int id) {
-        Customer customerToDelete = this.customerRepository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "No customer with that id was found")
-        );
+    public ResponseEntity<Response<?>> deleteCustomerById(@PathVariable int id) {
+        Customer customerToDelete = this.customerRepository.findById(id).orElse(null);
+        if (customerToDelete == null) {
+            this.errorResponse.set("No customer with that id found");
+            return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+        }
         this.customerRepository.delete(customerToDelete);
-        return ResponseEntity.ok(customerToDelete);
+        this.customerResponse.set(customerToDelete);
+        return ResponseEntity.ok(customerResponse);
     }
 
 }
